Separate state and stateless components
