+++ src/containers/DetailStockPage.js
