+++ src/containers/__test__/DetailStock.test.js
