+++ src/containers/__test__/SelectStockCategory.test.js
